requirements.txt
Flask==3.1.1
openai==1.82.1
chromadb==1.0.11
requests==2.32.3
GitPython==3.1.44
langchain==0.3.25
bootstrap-flask==2.5.0
python-dotenv==1.1.0 